+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added Tumblr sharer.
 
 ## [2.1.0] - 2018-10-04
 - Updated Twitter sharer url.